+++ src/game/Tactical/Interface_Utils.cc
@@ -90,7 +90,7 @@ void UnLoadCarPortraits(void)
 }
 
 
-static void DrawBar(UINT32 const XPos, UINT32 const YPos, UINT32 const Height, UINT32 const Color, UINT32 const ShadowColor, UINT16* const DestBuf)
+static void DrawBar(UINT32 const XPos, UINT32 const YPos, UINT32 const Height, UINT32 const Color, UINT32 const ShadowColor, UINT32* const DestBuf)
 {
 	LineDraw(TRUE, XPos + 0, YPos, XPos + 0, YPos - Height, ShadowColor, DestBuf);
 	LineDraw(TRUE, XPos + 1, YPos, XPos + 1, YPos - Height, Color,       DestBuf);
@@ -98,7 +98,7 @@ static void DrawBar(UINT32 const XPos, UINT32 const YPos, UINT32 const Height, U
 }
 
 
-static void DrawLifeUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 YPos, UINT32 const MaxHeight, UINT16* const pDestBuf)
+static void DrawLifeUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 YPos, UINT32 const MaxHeight, UINT32* const pDestBuf)
 {
 	UINT32 Height;
 
@@ -126,7 +126,7 @@ static void DrawLifeUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 YPos, 
 }
 
 
-static void DrawBreathUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 const sYPos, UINT32 const MaxHeight, UINT16* const pDestBuf)
+static void DrawBreathUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 const sYPos, UINT32 const MaxHeight, UINT32* const pDestBuf)
 {
 	UINT32 Height;
 
@@ -146,7 +146,7 @@ static void DrawBreathUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 cons
 }
 
 
-static void DrawMoraleUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 const YPos, UINT32 const MaxHeight, UINT16* const pDestBuf)
+static void DrawMoraleUIBar(SOLDIERTYPE const& s, UINT32 const XPos, UINT32 const YPos, UINT32 const MaxHeight, UINT32* const pDestBuf)
 {
 	UINT32 const Height = MaxHeight * s.bMorale / 100;
 	DrawBar(XPos, YPos, Height, MORALE_BAR, MORALE_BAR_SHADOW, pDestBuf);
@@ -189,7 +189,7 @@ void DrawSoldierUIBars(SOLDIERTYPE const& s, INT16 const sXPos, INT16 const sYPo
 
 	SGPVSurface::Lock l(uiBuffer);
 	SetClippingRegionAndImageWidth(l.Pitch(), 0, 0, SCREEN_WIDTH, SCREEN_HEIGHT);
-	UINT16* const pDestBuf = l.Buffer<UINT16>();
+	UINT32* const pDestBuf = l.Buffer<UINT32>();
 
 	DrawLifeUIBar(s, sXPos, sYPos, BarHeight, pDestBuf);
 	if (!(s.uiStatusFlags & SOLDIER_ROBOT))
@@ -228,7 +228,7 @@ void DrawItemUIBarEx(OBJECTTYPE const& o, const UINT8 ubStatus, const INT16 x, c
 
 	{ SGPVSurface::Lock l(uiBuffer);
 		SetClippingRegionAndImageWidth(l.Pitch(), 0, 0, SCREEN_WIDTH, SCREEN_HEIGHT);
-		UINT16* const pDestBuf = l.Buffer<UINT16>();
+		UINT32* const pDestBuf = l.Buffer<UINT32>();
 
 		--max_h; // LineDraw() includes the end point
 		const INT h = max_h * value / 100;
